#5965 - Importing CAS files with XML or PDF structure duplicates it
- Strip out XML/PDF structure from annotation CASes generated from generic CAS formats
